Complete reorganization
Update install.sh to work properly
Update autospeed internals to work properly
