+++ autospeed/__init__.py
@@ -1,0 +1,11 @@
+# Find your printers max speed before losing steps
+#
+# Copyright (C) 2024 Anonoei <dev@anonoei.com>
+#
+# This file may be distributed under the terms of the MIT license.
+
+from .funcs import *
+from .move import *
+from .wrappers import *
+
+from .main import AutoSpeed
