+++ autospeed/funcs.py
@@ -1,3 +1,9 @@
+# Find your printers max speed before losing steps
+#
+# Copyright (C) 2024 Anonoei <dev@anonoei.com>
+#
+# This file may be distributed under the terms of the MIT license.
+
 import math
 
 def calculate_velocity(accel: float, travel: float):
