+++ autospeed/wrappers.py
@@ -1,4 +1,10 @@
-from auto_speed_move import Move
+# Find your printers max speed before losing steps
+#
+# Copyright (C) 2024 Anonoei <dev@anonoei.com>
+#
+# This file may be distributed under the terms of the MIT license.
+
+from .move import Move
 
 class ResultsWrapper:
     def __init__(self):
